Add output_helpers.py
